unsaved-files: be safe when freeing an unsaved file
If the object has been copied the file and fd may be null and we
should not call unlink and close

+++ libide/ide-unsaved-files.c
@@ -75,9 +75,13 @@ unsaved_file_free (gpointer data)
     {
       g_object_unref (uf->file);
       g_bytes_unref (uf->content);
-      g_unlink (uf->temp_path);
-      g_free (uf->temp_path);
-      g_close (uf->temp_fd, NULL);
+      if (uf->temp_path)
+        {
+           g_unlink (uf->temp_path);
+           g_free (uf->temp_path);
+        }
+      if (uf->temp_fd != 0)
+        g_close (uf->temp_fd, NULL);
       g_slice_free (UnsavedFile, uf);
     }
 }
